refactor(frontend): organize patients feature by area
Splits features/patients into medicina/, nutricion/, and shared/ subfolders
to mirror @cais/shared/schemas structure. Components, hooks, sections, forms,
and constants specific to medicine move under medicina/; cross-area primitives
(StepFormShell, CIE10Field, formDefaults) move under shared/forms/.
