+++ frontend/src/features/patients/components/NewPatientButton.jsx
@@ -3,7 +3,7 @@ import Button from '@components/Button'
 import Modal from '@components/Modal'
 import usePermissions from '@hooks/usePermissions'
 import { AREAS } from '@cais/shared/constants/users'
-import MedicalPatientForm from '@features/patients/forms/MedicalPatientForm/MedicalPatientForm'
+import MedicalPatientForm from '@features/patients/medicina/forms/MedicalPatientForm/MedicalPatientForm'
 
 const AREA_FORMS = {
   [AREAS.MEDICINA]: {
