+++ frontend/src/features/patients/medicina/components/NoteCard.jsx
