+++ frontend/src/features/patients/medicina/components/NoteDetail.jsx
@@ -2,11 +2,11 @@ import { HiOutlineArrowLeft, HiOutlinePencilSquare } from 'react-icons/hi2'
 import Tab from '@components/Tab'
 import Button from '@components/Button'
 import DataField from '@components/DataField'
-import MotivoConsultaSection from '@features/patients/sections/MotivoConsultaSection'
-import PlanEstudioSection from '@features/patients/sections/PlanEstudioSection'
-import SignosVitalesSection from '@features/patients/sections/SignosVitalesSection'
-import FieldsSection from '@features/patients/sections/FieldsSection'
-import { buildAparSistFields } from '@features/patients/constants'
+import MotivoConsultaSection from '@features/patients/medicina/sections/MotivoConsultaSection'
+import PlanEstudioSection from '@features/patients/medicina/sections/PlanEstudioSection'
+import SignosVitalesSection from '@features/patients/medicina/sections/SignosVitalesSection'
+import FieldsSection from '@features/patients/medicina/sections/FieldsSection'
+import { buildAparSistFields } from '@features/patients/medicina/constants'
 import { formatFecha, formatHora } from '@lib/dateHelpers'
 
 export default function NoteDetail({ note, onBack, onEdit }) {
