+++ frontend/src/features/patients/medicina/forms/MedicalPatientForm/steps/AntecedentesNoPatStep.jsx
