+++ frontend/src/features/patients/medicina/forms/MedicalPatientForm/steps/AntecedentesPatologicosStep.jsx
