+++ frontend/src/features/patients/medicina/forms/MedicalPatientForm/steps/AparatosSistemasStep.jsx
