+++ frontend/src/features/patients/medicina/forms/MedicalPatientForm/steps/ExploracionFisicaStep.jsx
