+++ frontend/src/features/patients/medicina/hooks/useCreateMedicalHistory.js
