+++ frontend/src/features/patients/medicina/hooks/useCreatePatientWithHistory.js
