+++ frontend/src/features/patients/medicina/hooks/useMedicalHistories.js
