+++ frontend/src/features/patients/medicina/hooks/useMedicalHistory.js
