+++ frontend/src/features/patients/medicina/sections/CIE10List.jsx
