+++ frontend/src/features/patients/medicina/sections/Empty.jsx
