+++ frontend/src/features/patients/medicina/sections/InmunizacionesSection.jsx
