+++ frontend/src/features/patients/medicina/sections/MotivoConsultaSection.jsx
