+++ frontend/src/features/patients/medicina/sections/SignosVitalesSection.jsx
